moved ScalarSensor to regions/ here it belongs
from encoders/ where only ScalarEncoder should be
